pyproject.toml
[tool.poetry]
name = "frankenstein"
version = "0.1.0"
description = ""
authors = ["Haohang Li <hli113@stevens.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">= 3.10, < 3.11"
rich = "^13.7.0"
numpy = "^1.26.3"
pandas = "^2.1.4"
polars = "^0.20.3"
datasets = "^2.16.1"
accelerate = "^0.25.0"
dvc = {extras = ["s3"], version = "^3.39.0"}
black = {extras = ["jupyter"], version = "^23.12.1"}
torch = "^2.1.2"
transformers = {extras = ["torch"], version = "^4.36.2"}
typer = "^0.9.0"
ipykernel = "^6.28.0"
ipywidgets = "^8.1.1"
ujson = "^5.9.0"
hydra-core = "^1.3.2"
pillow = "^10.2.0"
matplotlib = "^3.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
